Se modifico el campo costo para poner un minimo, un máximo y aceptación de decimal

+++ database/migrations/2021_04_06_160152_create_calidad_plantas_table.php
@@ -27,7 +27,7 @@ class CreateCalidadPlantasTable extends Migration
             $table->string('resistenciaPlagas')->nullable();
             $table->string('toleranciaPlagas')->nullable();
             $table->string('certificado')->nullable();
-            $table->smallInteger('costo')->nullable();
+            $table->decimal('costo', 6, 1)->nullable();
             $table->foreignId('empleado_id')
                 ->references('id')
                 ->on('empleados');
