+++ database/migrations/2021_04_06_202331_create_registro_siembras_table.php
@@ -30,7 +30,7 @@ class CreateRegistroSiembrasTable extends Migration
             $table->float('distanciaPlanta',);
             $table->float('distanciaVesana',);
             $table->string('riego', 28);
-            $table->smallInteger('costo')->default(0);
+            $table->decimal('costo', 6, 1)->default(0);
             $table->foreignId('empleado_id')
             ->references('id')
             ->on('empleados');
